requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
httpx==0.25.1
pydantic==2.5.0
python-multipart==0.0.6
google-generativeai==0.3.1
pyyaml==6.0.1
